+++ widgetbook/lib/anyhoo_app_bar/anyhoo_bottom_bar.dart
@@ -8,9 +8,6 @@ import 'package:widgetbook_workspace/helpers/wrap_in_mocks_helper.dart';
 
 @widgetbook.UseCase(name: 'AnyhooBottomBar', type: AnyhooBottomBar, path: 'anyhoo_app_bar')
 Widget build(BuildContext context) {
-  final colorSchemeOptions = ['green', 'purple', 'red'];
-  final colorScheme = context.knobs.list(label: 'Color scheme', options: colorSchemeOptions, initialOption: 'purple');
-
   final backgroundColorOptions = {'none': null, 'brown': Colors.brown, 'orange': Colors.orange};
   final backgroundColor = context.knobs.list(
     label: 'Background color',
@@ -25,37 +22,24 @@ Widget build(BuildContext context) {
     AnyhooBottomBarItem(key: 'favorites', icon: Icons.favorite_outlined, label: 'Favorites', route: '/favorites'),
     AnyhooBottomBarItem(key: 'profile', icon: Icons.person_outlined, label: 'Profile', route: '/profile'),
   ];
-  final widget = Theme(
-    data: ThemeData(colorScheme: getColorScheme(colorScheme)),
-    child: Scaffold(
-      body: CustomScrollView(
-        controller: scrollController,
-        slivers: [
-          SliverList.builder(
-            itemCount: 100,
-            itemBuilder: (context, index) {
-              return Text('Item $index');
-            },
-          ),
-        ],
-      ),
-      bottomNavigationBar: AnyhooBottomBar(
-        backgroundColor: backgroundColorValue,
-        items: items,
-        selectedItemKey: items.first.key,
-      ),
+  final widget = Scaffold(
+    body: CustomScrollView(
+      controller: scrollController,
+      slivers: [
+        SliverList.builder(
+          itemCount: 100,
+          itemBuilder: (context, index) {
+            return Text('Item $index');
+          },
+        ),
+      ],
+    ),
+    bottomNavigationBar: AnyhooBottomBar(
+      backgroundColor: backgroundColorValue,
+      items: items,
+      selectedItemKey: items.first.key,
     ),
   );
 
   return WrapInMocksHelper().wrapInMocks(context, widget);
-}
-
-ColorScheme getColorScheme(String colorScheme) {
-  final cs = ColorScheme.fromSeed(seedColor: Colors.white);
-  return switch (colorScheme) {
-    'red' => cs.copyWith(primary: Colors.red),
-    'green' => cs.copyWith(primary: Colors.green),
-    'purple' => cs.copyWith(primary: Colors.purple),
-    _ => cs,
-  };
 }