+++ widgetbook/lib/anyhoo_image_selector/image_selector_widget.dart
@@ -9,9 +9,6 @@ import 'package:widgetbook_workspace/helpers/wrap_in_mocks_helper.dart';
 
 @widgetbook.UseCase(name: 'Default', type: AnyhooImageSelectorWidget, path: 'anyhoo_image_selector')
 Widget build(BuildContext context) {
-  final colorSchemeOptions = ['red', 'green', 'purple'];
-  final colorScheme = context.knobs.list(label: 'Color scheme', options: colorSchemeOptions, initialOption: 'red');
-
   final imageOptions = {'Asset': 'assets/images/baking.png', 'Web': 'https://picsum.photos/200/200', 'none': null};
   final imageSelection = context.knobs.list(
     label: 'Show image',
@@ -27,56 +24,43 @@ Widget build(BuildContext context) {
     initialOption: LayoutType.verticalStack.name,
   );
 
-  final widget = Theme(
-    data: ThemeData(colorScheme: getColorScheme(colorScheme)),
-    child: Center(
-      child: SizedBox(
-        width: 300,
-        height: 400,
-        child: Stack(
-          children: [
-            // Simulating how it might be used in a real app - inside a Stack
-            // This could be a form page with other widgets stacked on top
-            AnyhooImageSelectorWidget(
-              layoutType: LayoutType.values.firstWhere((e) => e.name == layoutType),
-              preselectedImage: image,
-              onImageSelected: (selectedImage) {
-                // ignore: avoid_print
-                print('selectedImage: $selectedImage');
-              },
-              roundImage: roundImage,
-              stockAssetPaths: [
-                'assets/images/baking.png',
-                'assets/images/casserole.png',
-                'assets/images/cocktail.png',
-                'assets/images/dessert.png',
-                'assets/images/drink.png',
-                'assets/images/pot.png',
-                'assets/images/stew.png',
-                'assets/images/baking.png',
-                'assets/images/casserole.png',
-                'assets/images/cocktail.png',
-                'assets/images/dessert.png',
-                'assets/images/drink.png',
-                'assets/images/pot.png',
-                'assets/images/stew.png',
-              ],
-            ),
-          ],
-        ),
+  final widget = Center(
+    child: SizedBox(
+      width: 300,
+      height: 400,
+      child: Stack(
+        children: [
+          // Simulating how it might be used in a real app - inside a Stack
+          // This could be a form page with other widgets stacked on top
+          AnyhooImageSelectorWidget(
+            layoutType: LayoutType.values.firstWhere((e) => e.name == layoutType),
+            preselectedImage: image,
+            onImageSelected: (selectedImage) {
+              // ignore: avoid_print
+              print('selectedImage: $selectedImage');
+            },
+            roundImage: roundImage,
+            stockAssetPaths: [
+              'assets/images/baking.png',
+              'assets/images/casserole.png',
+              'assets/images/cocktail.png',
+              'assets/images/dessert.png',
+              'assets/images/drink.png',
+              'assets/images/pot.png',
+              'assets/images/stew.png',
+              'assets/images/baking.png',
+              'assets/images/casserole.png',
+              'assets/images/cocktail.png',
+              'assets/images/dessert.png',
+              'assets/images/drink.png',
+              'assets/images/pot.png',
+              'assets/images/stew.png',
+            ],
+          ),
+        ],
       ),
     ),
   );
 
   return WrapInMocksHelper().wrapInMocks(context, widget);
-}
-
-ColorScheme getColorScheme(String colorScheme) {
-  final cs = ColorScheme.fromSeed(seedColor: Colors.white);
-  return switch (colorScheme) {
-    'red' => cs.copyWith(primary: Colors.red),
-    'green' => cs.copyWith(primary: Colors.green),
-    'purple' => cs.copyWith(primary: Colors.purple),
-    _ => cs,
-  };
 }