+++ widgetbook/lib/anyhoo_shimmer/shimmer.dart
@@ -8,31 +8,15 @@ import 'package:anyhoo_shimmer/anyhoo_shimmer.dart';
 
 @widgetbook.UseCase(name: 'Shimmer', type: AnyhooShimmer, path: 'anyhoo_shimmer')
 Widget build(BuildContext context) {
-  final colorSchemeOptions = ['green', 'purple', 'red'];
-  final colorScheme = context.knobs.list(label: 'Color scheme', options: colorSchemeOptions, initialOption: 'purple');
-
   final isLoading = context.knobs.boolean(label: 'Is loading', initialValue: true);
 
-  final widget = Theme(
-    data: ThemeData(colorScheme: getColorScheme(colorScheme)),
-    child: Center(
-      child: SizedBox(
-        width: 300,
-        height: 30,
-        child: AnyhooShimmer(child: isLoading ? Text('Hello world...') : ShimmerShapes.text()),
-      ),
+  final widget = Center(
+    child: SizedBox(
+      width: 300,
+      height: 30,
+      child: AnyhooShimmer(child: isLoading ? Text('Hello world...') : ShimmerShapes.text()),
     ),
   );
 
   return WrapInMocksHelper().wrapInMocks(context, widget);
-}
-
-ColorScheme getColorScheme(String colorScheme) {
-  final cs = ColorScheme.fromSeed(seedColor: Colors.white);
-  return switch (colorScheme) {
-    'red' => cs.copyWith(primary: Colors.red),
-    'green' => cs.copyWith(primary: Colors.green),
-    'purple' => cs.copyWith(primary: Colors.purple),
-    _ => cs,
-  };
 }