+++ widgetbook/lib/helpers/wrap_in_mocks_helper.dart
@@ -4,44 +4,24 @@ import 'dart:typed_data';
 import 'package:anyhoo_auth/cubit/anyhoo_auth_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-// import 'package:flutter_bloc/flutter_bloc.dart';
-// import 'package:widgetbook/widgetbook.dart';
 import 'package:widgetbook_workspace/helpers/device_frame_wrapper.dart';
-// import 'package:widgetbook_workspace/helpers/image_loader.dart';
 import 'package:widgetbook_workspace/helpers/mock_generator.dart';
 import 'package:image/image.dart' as img;
-
 import 'package:cross_file/cross_file.dart';
 
 class WrapInMocksHelper {
   Widget wrapInMocks(BuildContext context, Widget child) {
-    // final theme = Theme.of(context);
-
     final mockGenerator = MockGenerator();
 
-    // final isLoading = context.knobs.boolean(label: 'Show shimmer', initialValue: false);
-    final deviceFrameWrapper = DeviceFrameWrapper.wrapInDeviceFrame(context, child);
-
-    // return FutureBuilder(
-    //   future: ImageLoader().loadImages(),
-    //   builder: (context, snapshot) {
-    //     if (!snapshot.hasData) {
-    //       return Expanded(
-    //         child: Column(
-    //           mainAxisAlignment: MainAxisAlignment.center,
-    //           crossAxisAlignment: CrossAxisAlignment.center,
-    //           children: [CircularProgressIndicator()],
-    //         ),
-    //       );
-    //     }
-    //     final (imageFile, imageBytes) = snapshot.data!;
-
-    return MultiBlocProvider(
-      providers: [BlocProvider<AnyhooAuthCubit>(create: (context) => mockGenerator.getMockAuthCubit())],
-      child: deviceFrameWrapper,
+    final deviceFrameWrapper = DeviceFrameWrapper.wrapInDeviceFrame(
+      context,
+      MultiBlocProvider(
+        providers: [BlocProvider<AnyhooAuthCubit>(create: (context) => mockGenerator.getMockAuthCubit())],
+        child: child,
+      ),
     );
-    //   },
-    // );
+
+    return deviceFrameWrapper;
   }
 }